tests/wpt/web-platform-tests/html/webappapis/dynamic-markup-insertion/opening-the-input-stream/005.html

<!doctype html>
<title>Cancelling timeout after document.open</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id="log"></div>
<iframe src="/common/blank.html"></iframe>
<script>
var t = async_test();
var iframe;
onload = t.step_func(function() {
  var iframe = document.getElementsByTagName("iframe")[0];
  iframe.contentWindow.setTimeout(t.step_func(function() {assert_unreached()}), 100);
  assert_equals(iframe.contentDocument.open(), iframe.contentDocument);
  setTimeout(function() {t.done();}, 200);
});
</script>
